Removing deadwood
We are only interested in speeding up one branch of execution right
now. I've removed everything that doesn't have to do with checking all
the problems in a Grassmanian. This is a heavily simplified version of
the research code that will be what the next version in the next
language will be focused on.